refactor: extract version_tag and supporting test utils
